Export message definition types
These types are used in Webviz and are important now that MessageReader
does not call parseMessageDefinitions directly.
