+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rosbag",
-  "version": "2.6.0",
+  "version": "2.6.1",
   "license": "Apache-2.0",
   "repository": "cruise-automation/rosbag.js",
   "dependencies": {
